select: fix time-limit timeout for a pipe and a socket test failure
On some situtation, this test fails. For example,

    select-1 2>&1 | cat

This is because events are not cleared correctly after select() in
waitsocks(). So so_cancel() should be called first before clearing socket
events.

And ensure that waitsocks() thread ends. Otherwise, the thread may
be running after select() returns. This may lead to unexpected
situtations. For example, consecutive select() calls may be affected
by a previous select() call.

    modified:   network/select.c